fix(verify): split verify page into server/client components. Refs #187
Navbar uses next/headers (server-only). The verify page was "use client",
causing a build error. Split into server page.tsx (renders Navbar) and
client VerifyForm.tsx (handles form state/navigation).
